Initial structure of plugin
Add test function and tests for the function

Retab from spaces to tabs.

Correct return structure to match what repoir expects.

Remove unused dependencies and correct path.

Add prepublish script

If something does not exist we don't want to reject the promise.

We instead want to return it to the repoir as a problem.

Removing type that is only used in one spot.

Update README

Go with double quotes to avoid markdown highlighting

Fix JSON